pyproject.toml
[project]
name = "youtube-mcp"
version = "0.1.0"
description = "YouTube MCP server for video metadata, transcripts, and analysis"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.9.0",
    "google-api-python-client>=2.100.0",
    "youtube-transcript-api>=0.6.1",
    "python-dotenv>=1.0.0",
    "isodate>=0.6.1"
]

[project.scripts]
youtube-mcp = "youtube_mcp.server:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
